fixed small bug

+++ code/ParallelTesting.cpp
@@ -62,7 +62,7 @@ int main(int argc, char** argv) {
   for (int i = 0; i < 10; i++) {
     string output = "Iter " + to_string(i) + ", Thread " + to_string(omp_get_thread_num()) + "\n";
     cout << output;
-    for (int j = 0; j < 100000000000; j++) {
+    for (long j = 0; j < 100000000000; j++) {
       int k = i * i * i;
     }
   }
